// services/worker/src/protocols/smb/smb.protocol.type.ts
export enum SmbErrors {
    ACCESS_DENIED = 'NT_STATUS_ACCESS_DENIED',
    CONNECTION_REFUSED = 'NT_STATUS_CONNECTION_REFUSED',
    LOGON_FAILURE = 'NT_STATUS_LOGON_FAILURE',
    TIMEOUT = 'NT_STATUS_IO_TIMEOUT',
    PROTOCOL_MISMATCH = 'NT_STATUS_NOT_SUPPORTED',
    NETWORK_UNREACHABLE = 'NT_STATUS_NETWORK_UNREACHABLE',
    HOST_UNREACHABLE = 'NT_STATUS_HOST_UNREACHABLE',
    PORT_BLOCKED = 'NT_STATUS_PORT_BLOCKED'
}